+++ components/index.ts
@@ -1,5 +1,7 @@
-import { Button, ButtonLink, NavLink, Container } from './common';
+import { Button, ButtonLink, NavLink, Container, Modal } from './common';
+import { Dashboard } from './dashboard';
 import { Gallery, GalleryImage } from './gallery';
 import { MainNav } from './navigation';
 
-export { Button, ButtonLink, NavLink, Container, Gallery, GalleryImage, MainNav }
+
+export { Button, ButtonLink, NavLink, Container, Modal, Gallery, GalleryImage, MainNav, Dashboard }